requirements.txt
streamlit
pillow
pandas
numpy
scikit-learn
matplotlib
seaborn
